+++ scripts/upload-curseforge.ps1
@@ -54,6 +54,10 @@ function Get-Changelog {
         [string] $Path
     )
 
+    if ([string]::IsNullOrWhiteSpace($Path)) {
+        $Path = "changelogs/$Version/$Loader.md"
+    }
+
     if (![string]::IsNullOrWhiteSpace($Path)) {
         $resolvedPath = Join-Path $root $Path
         if (Test-Path -LiteralPath $resolvedPath) {